[feature-13] event.ts from Business layer was moved into results folder.
