+++ src/business/commands/i-create-event.ts
@@ -1,4 +1,4 @@
-import { Event } from '../models/event';
+import { Event } from '../models/results/event';
 import { IObservableCommand } from './core/i-observable-command';
 
 export interface ICreateEvent extends IObservableCommand<Event, Event> {
