--- src/business/models/event.ts
@@ -1,2 +1,0 @@
-export class Event {
-}
