+++ src/business/models/results/event.ts
@@ -1,0 +1,2 @@
+﻿export class Event {
+}
