+++ src/business/repositories/i-event-repository.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs';
-import { Event } from '../models/event';
+import { Event } from '../models/results/event';
 
 export interface IEventRepository {
     insert(event: Event): Observable<Event>;
